+++ api/src/test/kotlin/com/backgu/amaker/auth/test/SuccessfulStubGoogleApiClient.kt
@@ -1,4 +1,4 @@
-package com.backgu.amaker.auth.service
+package com.backgu.amaker.auth.test
 
 import com.backgu.amaker.auth.dto.GoogleUserInfoDto
 import com.backgu.amaker.auth.infra.GoogleApiClient
